Fixing bug in CDCL solver that had to do with not copying formulas
Why does it make sense to copy the formulas?
-> In case we back-jump the formula is already there at the node we jump
to. Otherwise we would need to reconstruct it when moving backwards.

The problem with the previous implementation was that back-jump was
actually never executed and we had no test (also not the random ones)
that discovered this. When adding a test-case where we need to backjump
(test11) the problem was present.