It seems like we have a problem with backjumping
We remove too many non decision literals sometimes. A solution would be
to keep another kind of edges in the graph that tell you which non
decision literal is there because of which decision literal. Then it
would be easier to remove the correct nodes in all cases.

Currently the problem only arises with large test-cases where we
back-jump multiple times...

- added test12.smt2 which is one of the larger cases where backjumping
  fails at some point...